backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pymongo>=4.5.0
motor>=3.3.0
pydantic>=2.4.2
python-dotenv>=1.0.0
stripe>=7.0.0
boto3>=1.28.0
cloudinary>=1.34.0
graphene>=3.3
aiosmtplib>=2.0.0
paypalrestsdk>=1.13.1
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
